flask_blog/requirements.txt
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
sqlparse==0.3.0
Werkzeug==0.15.4
WTForms==2.2.1
